+++ src/embind/embind_ts.js
@@ -31,8 +31,8 @@ var LibraryEmbind = {
     printSignature(nameMap, out) {
       out.push('(');
 
-      let argOut = [];
-      for (let arg of this.argumentTypes) {
+      const argOut = [];
+      for (const arg of this.argumentTypes) {
         argOut.push(`${arg.name}: ${nameMap(arg.type)}`);
       }
       out.push(argOut.join(', '));
@@ -67,7 +67,7 @@ var LibraryEmbind = {
         out.push(` extends ${this.base.name}`);
       }
       out.push(' {\n');
-      for (let method of this.methods) {
+      for (const method of this.methods) {
         out.push('  ');
         method.printFunction(nameMap, out);
       }
@@ -78,9 +78,85 @@ var LibraryEmbind = {
     printModuleEntry(nameMap, out) {
       out.push(`  ${this.name}: {new`);
       // TODO Handle constructor overloading
-      let constructor = this.constructors[this.constructors.length > 1 ? 1 : 0];
+      const constructor = this.constructors[this.constructors.length > 1 ? 1 : 0];
       constructor.printSignature(nameMap, out);
       out.push('};\n');
+    }
+  },
+  $ConstantDefinition: class ConstantDefinition {
+    constructor(type, name) {
+      this.type = type;
+      this.name = name;
+    }
+
+    printModuleEntry(nameMap, out) {
+      out.push(`  ${this.name}: ${nameMap(this.type)};\n`);
+    }
+  },
+  $EnumDefinition: class EnumDefinition {
+    constructor(typeId, name) {
+      this.typeId = typeId;
+      this.name = name;
+      this.items = [];
+    }
+
+    print(nameMap, out) {
+      out.push(`export interface ${this.name}Value<T extends number> {\n`);
+      out.push('  value: T;\n}\n');
+      out.push(`export type ${this.name} = `);
+      const outItems = [];
+      for (const [name, value] of this.items) {
+        outItems.push(`${this.name}Value<${value}>`);
+      }
+      out.push(outItems.join('|'));
+      out.push(';\n\n');
+    }
+
+    printModuleEntry(nameMap, out) {
+      out.push(`  ${this.name}: {`);
+      const outItems = [];
+      for (const [name, value] of this.items) {
+        outItems.push(`${name}: ${this.name}Value<${value}>`);
+      }
+      out.push(outItems.join(', '));
+      out.push('};\n');
+    }
+  },
+  $ValueArrayDefinition: class ValueArrayDefinition {
+    constructor(typeId, name) {
+      this.typeId = typeId;
+      this.name = name;
+      this.elementTypeIds = [];
+      this.elements = [];
+    }
+
+    print(nameMap, out) {
+      out.push(`export type ${this.name} = [ `);
+      const outElements = [];
+      for (const type of this.elements) {
+        outElements.push(nameMap(type));
+      }
+      out.push(outElements.join(', '))
+      out.push(' ];\n\n');
+    }
+  },
+  $ValueObjectDefinition: class ValueObjectDefinition {
+    constructor(typeId, name) {
+      this.typeId = typeId;
+      this.name = name;
+      this.fieldTypeIds = [];
+      this.fieldNames = [];
+      this.fields = [];
+    }
+
+    print(nameMap, out) {
+      out.push(`export type ${this.name} = {\n`);
+      const outFields = [];
+      for (const {name, type} of this.fields) {
+        outFields.push(`  ${name}: ${nameMap(type)}`);
+      }
+      out.push(outFields.join(',\n'))
+      out.push('\n};\n\n');
     }
   },
   $TsPrinter: class TsPrinter {
@@ -145,22 +221,22 @@ var LibraryEmbind = {
   },
   $createFunctionDefinition__deps: ['$FunctionDefinition', '$heap32VectorToArray', '$readLatin1String', '$Argument', '$whenDependentTypesAreResolved'],
   $createFunctionDefinition: function(name, argCount, rawArgTypesAddr, hasThis, cb) {
-    var argTypes = heap32VectorToArray(argCount, rawArgTypesAddr);
+    const argTypes = heap32VectorToArray(argCount, rawArgTypesAddr);
     name = readLatin1String(name);
 
     whenDependentTypesAreResolved([], argTypes, function(argTypes) {
-      var returnType = argTypes[0];
-      var thisType = null;
-      var argStart = 1;
+      const returnType = argTypes[0];
+      let thisType = null;
+      let argStart = 1;
       if (hasThis) {
         thisType = argTypes[1];
         argStart = 2;
       }
-      let args = [];
-      for (var i = argStart; i < argTypes.length; i++) {
+      const args = [];
+      for (let i = argStart; i < argTypes.length; i++) {
         args.push(new Argument(`_${i - argStart}`, argTypes[i]));
       }
-      var funcDef = new FunctionDefinition(name, returnType, args, thisType);
+      const funcDef = new FunctionDefinition(name, returnType, args, thisType);
       cb(funcDef);
       return [];
     });
@@ -222,7 +298,7 @@ var LibraryEmbind = {
       [rawType, rawPointerType, rawConstPointerType],
       baseClassRawType ? [baseClassRawType] : [],
       function(base) {
-        var classDef = new ClassDefinition(rawType, name, base.length ? base[0] : null);
+        const classDef = new ClassDefinition(rawType, name, base.length ? base[0] : null);
         moduleDefinitions.push(classDef);
         return [classDef, classDef, classDef];
       }
@@ -257,18 +333,144 @@ var LibraryEmbind = {
           isPureVirtual,
           isAsync) {
     createFunctionDefinition(methodName, argCount, rawArgTypesAddr, true, (funcDef) => {
-      let classDef = funcDef.thisType;
+      const classDef = funcDef.thisType;
       classDef.methods.push(funcDef);
+    });
+  },
+  _embind_register_enum__deps: ['$readLatin1String', '$EnumDefinition', '$moduleDefinitions'],
+  _embind_register_enum: function(rawType, name, size, isSigned) {
+    name = readLatin1String(name);
+    const enumDef = new EnumDefinition(rawType, name);
+    registerType(rawType, enumDef);
+    moduleDefinitions.push(enumDef);
+  },
+  _embind_register_enum_value__deps: ['$readLatin1String', '$requireRegisteredType'],
+  _embind_register_enum_value: function(rawEnumType, name, enumValue) {
+    name = readLatin1String(name);
+    const enumDef = requireRegisteredType(rawEnumType, name);
+    enumDef.items.push([name, enumValue]);
+  },
+  _embind_register_constant__deps: ['$readLatin1String', '$ConstantDefinition', '$whenDependentTypesAreResolved', '$moduleDefinitions'],
+  _embind_register_constant: function(name, typeId, value) {
+    name = readLatin1String(name);
+    whenDependentTypesAreResolved([], [typeId], function(types) {
+      const def = new ConstantDefinition(types[0], name);
+      moduleDefinitions.push(def);
+      return [];
+    });
+  },
+  _embind_register_value_array__deps: [
+    '$readLatin1String', '$ValueArrayDefinition', '$tupleRegistrations'],
+  _embind_register_value_array: function(
+    rawType,
+    name,
+    constructorSignature,
+    rawConstructor,
+    destructorSignature,
+    rawDestructor
+  ) {
+    name = readLatin1String(name);
+    const valueArray = new ValueArrayDefinition(rawType, name);
+    tupleRegistrations[rawType] = valueArray;
+  },
+  _embind_register_value_array_element__deps: ['$tupleRegistrations'],
+  _embind_register_value_array_element: function(
+    rawTupleType,
+    getterReturnType,
+    getterSignature,
+    getter,
+    getterContext,
+    setterArgumentType,
+    setterSignature,
+    setter,
+    setterContext
+  ) {
+    const valueArray = tupleRegistrations[rawTupleType];
+    assert(getterReturnType === setterArgumentType, 'Mismatched getter and setter types are not supported.');
+    valueArray.elementTypeIds.push(getterReturnType);
+  },
+  _embind_finalize_value_array__deps: ['$whenDependentTypesAreResolved', '$moduleDefinitions', '$tupleRegistrations'],
+  _embind_finalize_value_array: function(rawTupleType) {
+    const valueArray = tupleRegistrations[rawTupleType];
+    delete tupleRegistrations[rawTupleType];
+    whenDependentTypesAreResolved([rawTupleType], valueArray.elementTypeIds, function(types) {
+      moduleDefinitions.push(valueArray);
+      valueArray.elements = types;
+      return [valueArray];
+    });
+  },
+  _embind_register_value_object__deps: ['$readLatin1String', '$ValueObjectDefinition', '$structRegistrations'],
+  _embind_register_value_object: function(
+    rawType,
+    name,
+    constructorSignature,
+    rawConstructor,
+    destructorSignature,
+    rawDestructor
+  ) {
+    name = readLatin1String(name);
+    const valueObject = new ValueObjectDefinition(rawType, name);
+    structRegistrations[rawType] = valueObject;
+  },
+  _embind_register_value_object_field__deps: [
+    '$readLatin1String', '$structRegistrations'],
+  _embind_register_value_object_field: function(
+    structType,
+    fieldName,
+    getterReturnType,
+    getterSignature,
+    getter,
+    getterContext,
+    setterArgumentType,
+    setterSignature,
+    setter,
+    setterContext
+  ) {
+    const valueObject = structRegistrations[structType];
+    assert(getterReturnType === setterArgumentType, 'Mismatched getter and setter types are not supported.');
+    valueObject.fieldTypeIds.push(getterReturnType);
+    valueObject.fieldNames.push(readLatin1String(fieldName));
+  },
+  _embind_finalize_value_object__deps: ['$moduleDefinitions', '$whenDependentTypesAreResolved', '$structRegistrations'],
+  _embind_finalize_value_object: function(structType) {
+    const valueObject = structRegistrations[structType];
+    delete structRegistrations[structType];
+    whenDependentTypesAreResolved([structType], valueObject.fieldTypeIds, function(types) {
+      moduleDefinitions.push(valueObject);
+      for (let i = 0; i < types.length; i++) {
+        valueObject.fields.push({
+          name: valueObject.fieldNames[i],
+          type: types[i],
+        });
+      }
+      return [valueObject];
+    });
+  },
+  _embind_register_smart_ptr__deps: ['$whenDependentTypesAreResolved'],
+  _embind_register_smart_ptr: function(rawType,
+                                       rawPointeeType,
+                                       name,
+                                       sharingPolicy,
+                                       getPointeeSignature,
+                                       rawGetPointee,
+                                       constructorSignature,
+                                       rawConstructor,
+                                       shareSignature,
+                                       rawShare,
+                                       destructorSignature,
+                                       rawDestructor) {
+    whenDependentTypesAreResolved([rawType], [rawPointeeType], function(pointeeType) {
+      return [pointeeType[0]];
     });
   },
 
   $embindEmitTypes__deps: ['$awaitingDependencies', '$throwBindingError', '$getTypeName', '$moduleDefinitions', '$TsPrinter'],
   $embindEmitTypes__postset: 'addOnInit(embindEmitTypes);',
   $embindEmitTypes: function() {
-    for (var typeId in awaitingDependencies) {
+    for (const typeId in awaitingDependencies) {
       throwBindingError(`Missing type definition for '${getTypeName(typeId)}'`);
     }
-    var printer = new TsPrinter(moduleDefinitions);
+    const printer = new TsPrinter(moduleDefinitions);
     printer.print();
   },
 };